FAPI2 OP Tests: Handle Unregistered Redirect Uris Being Accpeted By OP.
RFC 9126 (PAR) section 2.4 states that an OP may accept a previously
unregistered 'redirect_uri'.

A new test, fapi2-security-profile-[id2|final]-ensure-registered-redirect-uri,
has been added to handle this case. This is applicable to baseline FAPI2 only.

If the authorization request is accepted at the PAR endpoint the test
proceeds to authentication and accessing a protected resource.