FAPI2 OP Tests: Handle Unregistered Redirect Uris Being Accepted By OP.
RFC 9126 (PAR) section 2.4 states that an OP may accept a previously
unregistered 'redirect_uri'.

A new test, fapi2-security-profile-[id2|final]-ensure-registered-redirect-uri,
has been added to handle this case. This is applicable to baseline FAPI2 only.

If the authorization request is accepted at the PAR endpoint the test
proceeds to authentication and accessing a protected resource.

A 'plain_fapi' run of 'fapi2-message-signing-final-test-plan' is added to the
CI tests to ensure the new test is run. The expected failures templates,
for existing tests, are updated for this new run.