Update DC API origin trial token
The last one expired in mid-April apparently. I think when it expires it
just causes an extra user confirmation step so it's not fatal, but here's
a new one that now expires 14 Oct 2025.

The tokens are created on https://developer.chrome.com/origintrials/#/trials/my
but the registrations are bound to a google account so the token is only
viewable when logged into my oidf google account.