OID4VCI: Add CI for issuer initiated variants
This is a little more complex than normal as it seems we need to wait for
the other test module to be setup before we do the redirect, so the
BrowserControl component is updated to have an optional delay before the
credential offer url is loaded.

I'm not sure this is actually the best approach, but it's the only one
I could manage to get working for now.

I think a longer term approach might be to make the issuer test not do
any retrieval of issuer metadata until it receives the credential offer,
and then we would be able to start that test module first and that start
the wallet test once the issuer test was ready to receive the credential
offer.