FAPI2 Final OP: Remove unnecessary MTLS key/cert/ca config fields
These fields are only necessary if using mtls client auth, mtls sender
constrained access tokens or the ecosystems requires tls to be used.
(i.e. they're not needed for plain_fapi + private_key_jwt + dpop)

Unfortunately we don't have a 'values' for @VariantConfigurationFields
so we have repeat the whole VariantConfigurationFields block for each
profile that needs mtls.

A similar change was made for VCI recently, although that case was a
little simpler.