FAPI2 Final RP Tests: Add Client Credentials Grant Test Plan
Add a new client test plan exclusively for the client credentials grant:
"fapi2-message-signing-final-client-credentials-grant-client-test-plan"

This is based on "fapi2-message-signing-final-client-test-plan" with
the 'id_token', 'authorization response', 'JARM', "Open Banking" and
'Refresh Token' tests removed.

testModulesWithVariants() in the new test plan defines a list of baseLine
variants. When added to the CI tests, as OP vs RP, the presence of the
variant causes run-test-plan.py to fail with an empty list of 'plan_modules'.

To address this we remove the addition of the variant to the test plan name
in the 'selected_modules' case allowing successful matches.