VPFinal: DCQL query schema
Update the DCQL schema to be a lot stricter.

Add checks that error if the DCQL query is invalid (missing required
fields etc), and warnings that warn if unknown properties are included.
(Unknown properties are allowed by the spec but the use of unknown
properties often happens because of a misunderstanding the specification.)

Rather than have two versions of the JSON schema, we achieve the
'warning for unknown fields' by having the schema reject unknown
fields, but filtering out those errors differently in the conditions
for warnings vs validation.

Unit tests and many changes generated by Claude & Codex.